Extract post-connections layout into re-usable class
