+++ assets/layout-list-detail.css
