+++ magic-link/templates/post-connections.php
@@ -36,7 +36,6 @@ class Disciple_Tools_Magic_Links_Template_Post_Connections extends DT_Magic_Url_
     public $record_post_type = 'groups'; // Child post type determined by the connection field selected
     private $post = null;
     private $items = [];
-    private $post_field_settings = null;
     private $meta_key = '';
 
     public $show_bulk_send = true;
@@ -52,6 +51,7 @@ class Disciple_Tools_Magic_Links_Template_Post_Connections extends DT_Magic_Url_
 
     private $template = null;
     private $link_obj = null;
+    private $layout = null;
 
     public function __construct( $template = null ) {
 
@@ -138,6 +138,12 @@ class Disciple_Tools_Magic_Links_Template_Post_Connections extends DT_Magic_Url_
         // Revert back to dt translations
         $this->hard_switch_to_default_dt_text_domain();
 
+        // Initialize layout front-end
+        $this->layout = new Disciple_Tools_Magic_Links_Layout_List_Detail(
+            $this->template,
+            $this->post,
+            $this->link_obj );
+
         /**
          * Load if valid url
          */
@@ -155,60 +161,19 @@ class Disciple_Tools_Magic_Links_Template_Post_Connections extends DT_Magic_Url_
     }
 
     public function wp_enqueue_scripts() {
-        $js_path = '../../assets/post-connections.js';
-        $css_path = '../../assets/post-connections.css';
-
-        wp_enqueue_style( 'ml-post-connections-css', plugin_dir_url( __FILE__ ) . $css_path, null, filemtime( plugin_dir_path( __FILE__ ) . $css_path ) );
-        wp_enqueue_script( 'ml-post-connections-js', plugin_dir_url( __FILE__ ) . $js_path, null, filemtime( plugin_dir_path( __FILE__ ) . $js_path ) );
-
-        $dtwc_version = '0.6.6';
-//        wp_enqueue_style( 'dt-web-components-css', "https://cdn.jsdelivr.net/npm/@disciple.tools/web-components@$dtwc_version/styles/light.css", [], $dtwc_version );
-        wp_enqueue_style( 'dt-web-components-css', "https://cdn.jsdelivr.net/npm/@disciple.tools/web-components@$dtwc_version/src/styles/light.css", [], $dtwc_version ); // remove 'src' after v0.7
-        wp_enqueue_script( 'dt-web-components-js', "https://cdn.jsdelivr.net/npm/@disciple.tools/web-components@$dtwc_version/dist/index.js", $dtwc_version );
-        add_filter( 'script_loader_tag', 'add_module_type_to_script', 10, 3 );
-        function add_module_type_to_script( $tag, $handle, $src ) {
-            if ( 'dt-web-components-js' === $handle ) {
-                // @codingStandardsIgnoreStart
-                $tag = '<script type="module" src="' . esc_url( $src ) . '"></script>';
-                // @codingStandardsIgnoreEnd
-            }
-            return $tag;
-        }
-        wp_enqueue_script( 'dt-web-components-services-js', "https://cdn.jsdelivr.net/npm/@disciple.tools/web-components@$dtwc_version/dist/services.min.js", array( 'jquery' ), true ); // not needed after v0.7
-
-        $mdi_version = '6.6.96';
-        wp_enqueue_style( 'material-font-icons-css', "https://cdn.jsdelivr.net/npm/@mdi/font@$mdi_version/css/materialdesignicons.min.css", [], $mdi_version );
+        $this->layout->wp_enqueue_scripts();
 
         Disciple_Tools_Bulk_Magic_Link_Sender_API::enqueue_magic_link_utilities_script();
     }
 
     public function dt_magic_url_base_allowed_js( $allowed_js ) {
-        $allowed_js[] = 'dt-web-components-js';
-        $allowed_js[] = 'dt-web-components-services-js';
-        $allowed_js[] = 'ml-post-connections-js';
         $allowed_js[] = Disciple_Tools_Bulk_Magic_Link_Sender_API::get_magic_link_utilities_script_handle();
 
-        return $allowed_js;
+        return $this->layout->allowed_js( $allowed_js );
     }
 
     public function dt_magic_url_base_allowed_css( $allowed_css ) {
-        $allowed_css[] = 'material-font-icons-css';
-        $allowed_css[] = 'dt-web-components-css';
-        $allowed_css[] = 'ml-post-connections-css';
-
-        return $allowed_css;
-    }
-
-    private function is_link_obj_field_enabled( $field_id ): bool {
-        if ( ! empty( $this->link_obj ) ) {
-            foreach ( $this->link_obj->type_fields ?? [] as $field ) {
-                if ( $field->id === $field_id ) {
-                    return $field->enabled;
-                }
-            }
-        }
-
-        return true;
+        return $this->layout->allowed_css( $allowed_css );
     }
 
     /**
@@ -217,214 +182,11 @@ class Disciple_Tools_Magic_Links_Template_Post_Connections extends DT_Magic_Url_
      * @see DT_Magic_Url_Base()->footer_javascript() for default state
      */
     public function footer_javascript() {
-        $localized_template_field_settings = DT_ML_Helper::localized_template_selected_field_settings( $this->template );
-        $localized_post_field_settings = DT_ML_Helper::localized_post_selected_field_settings( $this->post, $localized_template_field_settings, [ 'ID', 'post_type' ] );
-        ?>
-        <script>
-            let jsObject = [<?php echo json_encode( [
-                'root'                    => esc_url_raw( rest_url() ),
-                'nonce'                   => wp_create_nonce( 'wp_rest' ),
-                'parts'                   => $this->parts,
-                'post'                    => $localized_post_field_settings,
-                'items'                   => $this->items,
-                'template'                => $this->template,
-                'fieldSettings' => $localized_template_field_settings, //todo: should be for sub-type
-                'translations'            => [
-                    'regions_of_focus' => __( 'Regions of Focus', 'disciple_tools' ),
-                    'all_locations'    => __( 'All Locations', 'disciple_tools' ),
-                    'update_success' => __( 'Update Successful!', 'disciple_tools' ),
-                    'validation' => [
-                        'number' => [
-                            'out_of_range' => __( 'Value out of range!', 'disciple_tools' )
-                        ]
-                    ]
-                ],
-                'mapbox'                  => [
-                    'map_key'        => DT_Mapbox_API::get_key(),
-                    'google_map_key' => Disciple_Tools_Google_Geocode_API::get_key(),
-                    'translations'   => [
-                        'search_location' => __( 'Search Location', 'disciple_tools' ),
-                        'delete_location' => __( 'Delete Location', 'disciple_tools' ),
-                        'use'             => __( 'Use', 'disciple_tools' ),
-                        'open_modal'      => __( 'Open Modal', 'disciple_tools' )
-                    ]
-                ]
-            ] ) ?>][0];
-
-            const listItems = new Map(jsObject.items.posts.map((obj) => [obj.ID.toString(), obj]));
-
-            // initialize the list of items
-            loadListItems();
-        </script>
-        <?php
-        return true;
+        $this->layout->footer_javascript( $this->parts, $this->items );
     }
 
     public function body() {
-        $has_title = ! empty( $this->template ) && ( isset( $this->template['title'] ) && ! empty( $this->template['title'] ) );
-        ?>
-        <main>
-            <div id="list" class="is-expanded">
-                <header>
-                    <h1><?php echo $has_title ? esc_html( $this->template['title'] ) : '&nbsp;' ?></h1>
-                    <button type="button" class="mdi mdi-web" onclick="document.getElementById('post-locale-modal')._openModal()"></button>
-                    <button type="button" class="mdi mdi-information-outline" onclick="document.getElementById('post-detail-modal')._openModal()"></button>
-                </header>
-                <div id="search-filter">
-                    <div id="search-bar">
-                    <input type="text" id="search" placeholder="Search" onkeyup="searchChange(<?php echo esc_html( $this->post['ID'] ) ?>)" />
-                    <button id="clear-button" style="display: none;" class="clear-button mdi mdi-close" onclick="clearSearch(<?php echo esc_html( $this->post['ID'] ) ?>)"></button>
-                        <button class="filter-button mdi mdi-filter-variant" onclick="toggleFilters()"></button>
-
-                    </div>
-                    <div class="filters hidden">
-                        <div class="container">
-                            <h3>Sort By</h3>
-                            <label>
-                                <input type="radio" name="sort" value="-updated_at" onclick="toggleFilters()" onchange="searchChange(<?php echo esc_html( $this->post['ID'] ) ?>)" checked />
-                                Last Updated
-                            </label>
-                            <label>
-                                <input type="radio" name="sort" value="name" onclick="toggleFilters()" onchange="searchChange(<?php echo esc_html( $this->post['ID'] ) ?>)" />
-                                Name (A-Z)
-                            </label>
-                            <label>
-                                <input type="radio" name="sort" value="-name" onclick="toggleFilters()" onchange="searchChange(<?php echo esc_html( $this->post['ID'] ) ?>)" />
-                                Name (Z-A)
-                            </label>
-                        </div>
-                    </div>
-                </div>
-                <ul id="list-items" class="items"></ul>
-                <div id="spinner-div" style="justify-content: center; display: flex;">
-                    <span id="temp-spinner" class="loading-spinner inactive" style="margin: 0; position: absolute; top: 50%; -ms-transform: translateY(-50%); transform: translateY(-50%); height: 100px; width: 100px; z-index: 100;"></span>
-                </div>
-                <template id="list-item-template">
-                    <li>
-                        <a href="javascript:loadPostDetail()">
-                            <span class="post-id"></span>
-                            <span class="post-title"></span>
-                            <span class="post-updated-date"></span>
-                        </a>
-                    </li>
-                </template>
-            </div>
-            <div id="detail" class="">
-                <!-- <form onsubmit="saveItem(event)"> -->
-                <form>
-                    <header>
-                        <button type="button" class="details-toggle mdi mdi-arrow-left" onclick="togglePanels()"></button>
-                        <h2 id="detail-title"></h2>
-                        <span id="detail-title-post-id"></span>
-                    </header>
-
-                    <div id="detail-content"></div>
-                    <footer>
-                        <dt-button onclick="saveItem(event)" type="submit" context="primary"><?php esc_html_e( 'Submit Update', 'disciple_tools' ) ?></dt-button>
-                    </footer>
-                </form>
-
-                <template id="comment-header-template">
-                    <div class="comment-header">
-                        <span><strong id="comment-author"></strong></span>
-                        <span class="comment-date" id="comment-date"></span>
-                    </div>
-                </template>
-                <template id="comment-content-template">
-                    <div class="activity-text">
-                        <div dir="auto" class="" data-comment-id="" id="comment-id">
-                            <div class="comment-text" title="" dir="auto" id="comment-content">
-                            </div>
-                        </div>
-                    </div>
-                </template>
-
-                <template id="post-detail-template">
-                    <input type="hidden" name="id" id="post-id" />
-                    <input type="hidden" name="type" id="post-type" />
-
-                    <dt-tile id="all-fields" open>
-                    <?php
-                    $post_field_settings = DT_Posts::get_post_field_settings( $this->template['record_type'] );
-                    if ( isset( $this->template['fields'] ) ) {
-                        foreach ( $this->template['fields'] as $field ) {
-                            if ( !$field['enabled'] || !$this->is_link_obj_field_enabled( $field['id'] ) ) {
-                                continue;
-                            }
-
-                            if ( $field['type'] === 'dt' ) {
-                                // display standard DT fields
-                                $post_field_settings[$field['id']]['custom_display'] = false;
-                                $post_field_settings[$field['id']]['readonly'] = !empty( $field['readonly'] ) && boolval( $field['readonly'] );
-
-                                Disciple_Tools_Magic_Links_Helper::render_field_for_display( $field['id'], $post_field_settings, [] );
-                            } else {
-                                // display custom field for this magic link
-                                $tag = isset( $field['custom_form_field_type'] ) && $field['custom_form_field_type'] == 'textarea'
-                                    ? 'dt-textarea'
-                                    : 'dt-text';
-                                $label = ( ! empty( $field['translations'] ) && isset( $field['translations'][ determine_locale() ] ) ) ? $field['translations'][ determine_locale() ]['translation'] : $field['label'];
-                                ?>
-                                <<?php echo esc_html( $tag ) ?>
-                                    id="<?php echo esc_html( $field['id'] ) ?>"
-                                    name="<?php echo esc_html( $field['id'] ) ?>"
-                                    data-type="<?php echo esc_attr( $field['type'] ) ?>"
-                                    label="<?php echo esc_attr( $label ) ?>"
-                                ></<?php echo esc_html( $tag ) ?>>
-                                <?php
-                            }
-                        }
-                    }
-                    ?>
-                    </dt-tile>
-
-                    <dt-tile id="comments-tile" title="Comments">
-                        <div>
-                            <textarea id="comments-text-area"
-                                      style="resize: none;"
-                                      placeholder="<?php echo esc_html_x( 'Write your comment or note here', 'input field placeholder', 'disciple_tools' ) ?>"
-                            ></textarea>
-                        </div>
-                        <div class="comment-button-container">
-                            <button class="button loader" type="button" id="comment-button">
-                                <?php esc_html_e( 'Submit comment', 'disciple_tools' ) ?>
-                            </button>
-                        </div>
-                    </dt-tile>
-                </template>
-            </div>
-            <div id="snackbar-area"></div>
-            <template id="snackbar-item-template">
-                <div class="snackbar-item"></div>
-            </template>
-        </main>
-        <dt-modal id="post-detail-modal" buttonlabel="Open Modal" hideheader hidebutton closebutton>
-            <span slot="content" id="post-detail-modal-content">
-                <span class="post-name"><?php echo esc_html( $this->post['name'] ) ?></span>
-                <span class="post-id">ID: <?php echo esc_html( $this->post['ID'] ) ?></span>
-            </span>
-        </dt-modal>
-        <?php
-        $lang = dt_get_available_languages();
-        $current_lang = trim( wp_get_current_user()->locale );
-        ?>
-        <dt-modal id="post-locale-modal" buttonlabel="Open Modal" hideheader hidebutton closebutton>
-            <span slot="content" id="post-locale-modal-content">
-            <ul class="language-select">
-                <?php
-                foreach ( $lang as $language ) {
-                    ?>
-                    <li
-                        class="<?php echo $language['language'] === $current_lang ? esc_attr( 'active' ) : null ?>"
-                        onclick="assignLanguage('<?php echo esc_html( $language['language'] ); ?>')"
-                    ><?php echo esc_html( $language['native_name'] ); ?></li>
-                    <?php
-                }
-                ?>
-                </ul>
-            </span>
-        </dt-modal>
-        <?php
+        $this->layout->body();
     }
 
     /**
